libxtreemfs: jni: Move native libxtreemfs to the servers directory.
