--- java/jni/.gitignore
@@ -1,7 +1,0 @@
-.classpath
-.project
-*.class
-/src/org/xtreemfs/common/libxtreemfs/swig/
-/build/
-/dist/
-/.settings/
